Samples: AtomiCounters - finish the sample and drop obsolete code
also add README how to generate thumbnails as I keep forgetting it
myself.
