--- Samples/Media/materials/programs/GLSL150/AtomicVS.glsl
@@ -1,11 +1,0 @@
-#version 150
-
-in vec4 vertex;
-
-uniform mat4 WorldViewProj;
-
-void main()
-{
-    //gl_Position = WorldViewProj * vertex;
-    gl_Position = vertex;
-}
